follow pattern of i18n js for formValidation
